pyproject.toml
[tool.poetry]
name = "betstgbot"
version = "0.1.0"
description = ""
authors = ["Egor <kogayegor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiogram = "^3.10.0"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
